Add more method definitions -- see description
Some of them need to be reworked (as described in L3-5) but I wanted
to add them to the object for now. I'll work through ensuring the input
of each is validated and handled properly moving forward.

A complete list of where these methods are derived from can be found here:
https://minecraft.gamepedia.com/Raw_JSON_text_format#Formatting